move to utils.cpp
